requirements.txt
requests==2.22.0
dataclasses==0.7
pyyaml==5.1.2
shapely[vectorized]==1.6.4.post2
marshmallow==3.2.1
